Added peak finding, optimized fc.plot_behavior and updated the functions used to save csv data and plot pdfs
Added a peak finding algorithm that uses scipy.signal.find_peaks_cwt.

Updated behavior plotting so that the starts and stops for each behavior is only found once per object instead of once per channel.

Changed the output of fc.plot_behavior to be a list of plots instead of one plot with multiple subplots.