Added finding peak results and fixed bug in finding peaks
I added a dataframe to fold results of finding_peaks called peak_results.

Though this doesn't exist in earlier versions it should not cause issues with fibers objects that were created in previous versions.
I also fixed a couple bugs in peak_finding. There was an indexing issue that cause problems when the start time was not 0. I also did a quick adjustments to the peaks so that it found the actual highest point at each peak.