Updated Normalization function
Now if the biexp threshold is 1 the code will not attempt to fit a biexponential to the signal or reference. This should avoid any runtime errors when the signal does not have a decay and make the code run faster for those who do not wish to linearize their traces.